move `glossary` to `/data`
