+++ apps/frontend/src/data/glossary.ts
