+++ apps/frontend/src/routes/pokemon-roulette.lazy.tsx
@@ -2,7 +2,7 @@ import { createLazyFileRoute } from '@tanstack/react-router'
 import Gallery from '@/components/Gallery'
 import ProjectsNav from '@/components/ProjectsNav'
 import { game, loss, guess, win } from '@/assets/pokemon-roulette'
-import { typeScript, vue, vite, pinia, pokeApi } from '@/utilities/glossary'
+import { typeScript, vue, vite, pinia, pokeApi } from '@/data/glossary'
 
 export const Route = createLazyFileRoute('/pokemon-roulette')({
     component: PokemonRoulette,
